backend: update / reorganize remaining testing
